Created a cards folder. Moved ProjectCard out of Applications into its own file and used it in ProjectManagementDialog
