+++ frontend/src/components/cards/HowItWorksCard.tsx
